+++ packages/rich-text-links/src/index.ts
@@ -3,11 +3,13 @@ import {
   BLOCKS,
   Document,
   Inline,
+  INLINES,
   Link,
   Node,
   NodeData,
   ResourceLink,
 } from '@contentful/rich-text-types';
+import { Maybe } from './types/utils';
 
 export type EntityLinks = { [type in EntityType]: EntityLink[] };
 export type EntityLinkMaps = { [type in EntityType]: Map<string, EntityLink> };
@@ -19,22 +21,46 @@ export type EntityLinkNodeData = {
 };
 
 // spare upstream dependencies the need to use rich-text-types
-type AcceptedResourceLinkTypes = `${BLOCKS.EMBEDDED_RESOURCE}`;
-// | `${INLINES.EMBEDDED_RESOURCE}`
-// | `${INLINES.RESOURCE_HYPERLINK}`;
+type AcceptedResourceLinkTypes =
+  | `${BLOCKS.EMBEDDED_RESOURCE}`
+  | `${INLINES.EMBEDDED_RESOURCE}`
+  | `${INLINES.RESOURCE_HYPERLINK}`;
 
 /**
  * Extracts all links no matter the entity they are pointing to.
  */
 export function getRichTextResourceLinks(
-  document: Document,
+  document: Maybe<Document>,
   nodeType: AcceptedResourceLinkTypes,
   { deduplicate = true }: { deduplicate?: boolean } = {},
 ): ResourceLink[] {
-  const links = new Map<string, ResourceLink>();
   const isValidType = (actualNodeType: string, data: NodeData) =>
     actualNodeType === nodeType && data.target?.sys?.type === 'ResourceLink';
 
+  return getValidResourceLinks(document, isValidType, deduplicate);
+}
+
+export function getAllRichTextResourceLinks(
+  document: Maybe<Document>,
+  { deduplicate = true }: { deduplicate?: boolean } = {},
+): ResourceLink[] {
+  const nodeTypes: string[] = [
+    BLOCKS.EMBEDDED_RESOURCE,
+    INLINES.EMBEDDED_RESOURCE,
+    INLINES.RESOURCE_HYPERLINK,
+  ];
+  const isValidType = (actualNodeType: string, data: NodeData) =>
+    nodeTypes.includes(actualNodeType) && data.target?.sys?.type === 'ResourceLink';
+
+  return getValidResourceLinks(document, isValidType, deduplicate);
+}
+
+function getValidResourceLinks(
+  document: Maybe<Document>,
+  isValidType: (actualNodeType: string, data: NodeData) => boolean,
+  deduplicate: boolean,
+): ResourceLink[] {
+  const links = new Map<string, ResourceLink>();
   visitNodes(document, (node) => {
     if (isValidType(node.nodeType, node.data)) {
       const key = deduplicate ? node.data.target.sys.urn : links.size;
@@ -52,7 +78,7 @@ export function getRichTextEntityLinks(
   /**
    *  An instance of a Rich Text Document.
    */
-  document: Document,
+  document: Maybe<Document>,
   /**
    *  Node type. Only the entity links with given node type will be extracted.
    */
@@ -84,7 +110,7 @@ function isContentNode(node: Node): node is Inline | Block {
   return 'content' in node && Array.isArray(node.content);
 }
 
-function visitNodes(startNode: Node, onVisit: (node: Node) => void): void {
+function visitNodes(startNode: Maybe<Node>, onVisit: (node: Node) => void): void {
   /**
    * Do not remove this null check as consumers of this library do not all have good Typescript types
    */