+++ packages/rich-text-plain-text-renderer/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [16.0.11](https://github.com/contentful/rich-text/compare/@contentful/rich-text-plain-text-renderer@16.0.10...@contentful/rich-text-plain-text-renderer@16.0.11) (2024-01-30)
+
+**Note:** Version bump only for package @contentful/rich-text-plain-text-renderer
+
+## [16.0.10](https://github.com/contentful/rich-text/compare/@contentful/rich-text-plain-text-renderer@16.0.9...@contentful/rich-text-plain-text-renderer@16.0.10) (2024-01-23)
+
+**Note:** Version bump only for package @contentful/rich-text-plain-text-renderer
+
+## [16.0.9](https://github.com/contentful/rich-text/compare/@contentful/rich-text-plain-text-renderer@16.0.8...@contentful/rich-text-plain-text-renderer@16.0.9) (2024-01-23)
+
+**Note:** Version bump only for package @contentful/rich-text-plain-text-renderer
+
+## [16.0.8](https://github.com/contentful/rich-text/compare/@contentful/rich-text-plain-text-renderer@16.0.7...@contentful/rich-text-plain-text-renderer@16.0.8) (2024-01-23)
+
+**Note:** Version bump only for package @contentful/rich-text-plain-text-renderer
+
+## [16.0.7](https://github.com/contentful/rich-text/compare/@contentful/rich-text-plain-text-renderer@16.0.6...@contentful/rich-text-plain-text-renderer@16.0.7) (2023-09-12)
+
+**Note:** Version bump only for package @contentful/rich-text-plain-text-renderer
+
+## [16.0.6](https://github.com/contentful/rich-text/compare/@contentful/rich-text-plain-text-renderer@16.0.5...@contentful/rich-text-plain-text-renderer@16.0.6) (2023-08-04)
+
+**Note:** Version bump only for package @contentful/rich-text-plain-text-renderer
+
 ## [16.0.5](https://github.com/contentful/rich-text/compare/@contentful/rich-text-plain-text-renderer@16.0.4...@contentful/rich-text-plain-text-renderer@16.0.5) (2023-05-26)
 
 **Note:** Version bump only for package @contentful/rich-text-plain-text-renderer