+++ packages/rich-text-types/CHANGELOG.md
@@ -3,6 +3,32 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [16.3.4](https://github.com/contentful/rich-text/compare/@contentful/rich-text-types@16.3.3...@contentful/rich-text-types@16.3.4) (2024-01-30)
+
+**Note:** Version bump only for package @contentful/rich-text-types
+
+## [16.3.3](https://github.com/contentful/rich-text/compare/@contentful/rich-text-types@16.3.2...@contentful/rich-text-types@16.3.3) (2024-01-23)
+
+**Note:** Version bump only for package @contentful/rich-text-types
+
+## [16.3.2](https://github.com/contentful/rich-text/compare/@contentful/rich-text-types@16.3.1...@contentful/rich-text-types@16.3.2) (2024-01-23)
+
+**Note:** Version bump only for package @contentful/rich-text-types
+
+## [16.3.1](https://github.com/contentful/rich-text/compare/@contentful/rich-text-types@16.3.0...@contentful/rich-text-types@16.3.1) (2024-01-23)
+
+**Note:** Version bump only for package @contentful/rich-text-types
+
+# [16.3.0](https://github.com/contentful/rich-text/compare/@contentful/rich-text-types@16.2.1...@contentful/rich-text-types@16.3.0) (2023-09-12)
+
+### Features
+
+- add new nodes [DANTE-1157] ([#494](https://github.com/contentful/rich-text/issues/494)) ([b1fa6df](https://github.com/contentful/rich-text/commit/b1fa6dffc4bd7e1d367e9ce2cfddffe4fe07be47))
+
+## [16.2.1](https://github.com/contentful/rich-text/compare/@contentful/rich-text-types@16.2.0...@contentful/rich-text-types@16.2.1) (2023-08-04)
+
+**Note:** Version bump only for package @contentful/rich-text-types
+
 # [16.2.0](https://github.com/contentful/rich-text/compare/@contentful/rich-text-types@16.1.0...@contentful/rich-text-types@16.2.0) (2023-05-26)
 
 ### Features